fix: address review feedback — shared retention helper, default cutoff, bounded cache
- Extract checkRetention/getRetentionCutoff into shared lib/server/retention.ts
  (eliminates duplication between detector-counts and runs proxies)
- Runs proxy: apply default retention cutoff when start_after is omitted,
  so unbounded queries stay within the access window (P1)
- has_traces cache: replace infinite TTL with 1-hour TTL and cap at 1024
  entries to prevent unbounded memory growth (P2)
- Sessions test docstring: update to match enforce behavior (P2)
