+++ frontend/ui/src/app/api/projects/[projectId]/detector-counts/route.ts
@@ -1,39 +1,11 @@
-import { NextRequest, NextResponse } from "next/server";
+import { NextRequest } from "next/server";
 import { requireAuth, requireProjectAccess, errorResponse } from "@/lib/auth-helpers";
 import { prisma, PlanType } from "@traceroot/core";
+import { checkRetention } from "@/lib/server/retention";
 import { env } from "@/env";
 
 const BACKEND_URL = process.env.BACKEND_INTERNAL_URL || "http://localhost:8000";
 const INTERNAL_API_SECRET = env.INTERNAL_API_SECRET || "";
-
-const PLAN_RETENTION_DAYS: Record<string, number | null> = {
-  [PlanType.FREE]: 15,
-  [PlanType.STARTER]: 30,
-  [PlanType.PRO]: 90,
-  [PlanType.ENTERPRISE]: null,
-};
-const FAIL_CLOSED_DAYS = 15;
-
-function checkRetention(billingPlan: string, startAfter: string): NextResponse | null {
-  const days =
-    billingPlan in PLAN_RETENTION_DAYS ? PLAN_RETENTION_DAYS[billingPlan] : FAIL_CLOSED_DAYS;
-  if (days === null) return null;
-  const cutoff = new Date(Date.now() - days * 86_400_000 - 3_600_000);
-  if (new Date(startAfter) < cutoff) {
-    return NextResponse.json(
-      {
-        detail: {
-          message: "Data outside retention window",
-          retention_days: days,
-          cutoff: cutoff.toISOString(),
-          plan: billingPlan,
-        },
-      },
-      { status: 403 },
-    );
-  }
-  return null;
-}
 
 type RouteParams = { params: Promise<{ projectId: string }> };
